📝 Added missing files to the documentation
Multiple files have not been added to `mkdocs` which is resolved by this commit.
