+++ aswe/core/agent.py
@@ -12,7 +12,7 @@ from pandas.errors import IndexingError
 from aswe.core.data import BestMatch, User
 from aswe.core.use_cases import EventUseCase, GeneralUseCase, TransportationUseCase
 from aswe.core.user_interaction import SpeechToText, TextToSpeech
-from aswe.utils.general import clear_shell
+from aswe.utils.shell import clear_shell
 
 
 class Agent:
