+++ aswe/core/use_cases/transportation.py
@@ -7,7 +7,7 @@ from aswe.api.calendar.calendar import get_next_event_today
 from aswe.api.weather.weather import WeatherApi
 from aswe.api.weather.weather_params import DynamicPeriodEnum, ElementsEnum, IncludeEnum
 from aswe.core.data import BestMatch
-from aswe.utils.classes import AbstractUseCase
+from aswe.utils.abstract import AbstractUseCase
 
 
 class TransportationUseCase(AbstractUseCase):
