+++ aswe/utils/abstract.py
