+++ aswe/utils/shell.py
@@ -1,0 +1,10 @@
+import os
+from sys import platform
+
+
+def clear_shell() -> None:
+    """Clears any previous text in the shell"""
+    if platform == "linux" or platform == "linux2" or platform == "darwin":
+        os.system("clear")
+    elif platform == "win32":
+        os.system("cls")
