+++ docs/api/news.md
@@ -1,0 +1,3 @@
+# News
+
+::: aswe.api.news.news
