+++ docs/api/sport.md
@@ -1,0 +1,9 @@
+# Sport
+
+::: aswe.api.sport.basketball
+
+::: aswe.api.sport.f1
+
+::: aswe.api.sport.football
+
+::: aswe.api.sport.handball
