--- docs/core/use_cases.md
@@ -1,3 +1,0 @@
-# Use Cases
-
-::: aswe.core.use_case
