+++ docs/core/use_cases/event.md
@@ -1,0 +1,3 @@
+# Event
+
+::: aswe.core.use_cases.event
