+++ docs/core/use_cases/general.md
@@ -1,0 +1,3 @@
+# General
+
+::: aswe.core.use_cases.general
