+++ docs/core/use_cases/morning_briefing.md
@@ -1,0 +1,3 @@
+# Morning Briefing
+
+::: aswe.core.use_cases.morning_briefing
