+++ docs/core/use_cases/sport.md
@@ -1,0 +1,3 @@
+# Sport
+
+::: aswe.core.use_cases.sport
