+++ docs/core/use_cases/transportation.md
@@ -1,0 +1,3 @@
+# Transportation
+
+::: aswe.core.use_cases.transportation
