+++ docs/utils/abstract.md
@@ -1,0 +1,3 @@
+# Abstract Classes
+
+::: aswe.utils.abstract
