+++ docs/utils/date.md
@@ -1,0 +1,3 @@
+# Date
+
+::: aswe.utils.date
