+++ docs/utils/error.md
@@ -1,0 +1,3 @@
+# Error
+
+::: aswe.utils.error
