+++ docs/utils/request.md
@@ -1,0 +1,3 @@
+# Requests
+
+::: aswe.utils.request
