+++ docs/utils/shell.md
@@ -1,0 +1,3 @@
+# Shell
+
+::: aswe.utils.shell
