--- docs/utils/utils.md
@@ -1,7 +1,0 @@
-# Utilities
-
-::: aswe.utils.general
-
-::: aswe.utils.validate
-
-::: aswe.utils.request
