+++ docs/utils/validate.md
@@ -1,0 +1,3 @@
+# Validate
+
+::: aswe.utils.validate
