+++ tests/api/sports/test_basketball.py
@@ -1,6 +1,6 @@
 import pytest
 
-from aswe.api.sports.basketball import (
+from aswe.api.sport.basketball import (
     get_nba_standings,
     get_team_game_today,
     get_team_id,
@@ -10,7 +10,7 @@ from aswe.utils.error import ApiLimitReached
 
 @pytest.mark.xfail(raises=ApiLimitReached)
 def test_get_nba_standings() -> None:
-    """Test `aswe.api.sports.basketball.get_nba_standings`"""
+    """Test `aswe.api.sport.basketball.get_nba_standings`"""
     standings = get_nba_standings()
 
     assert len(standings) == 2 and len(standings[0]) == 15 and len(standings[1]) == 15
@@ -18,7 +18,7 @@ def test_get_nba_standings() -> None:
 
 @pytest.mark.xfail(raises=ApiLimitReached)
 def test_get_team_game_today() -> None:
-    """Test `aswe.api.sports.basketball.get_team_game_today`"""
+    """Test `aswe.api.sport.basketball.get_team_game_today`"""
     game = get_team_game_today("Boston Celtics")
 
     assert isinstance(game, list)
@@ -30,7 +30,7 @@ def test_get_team_game_today() -> None:
 
 @pytest.mark.xfail(raises=ApiLimitReached)
 def test_get_team_id() -> None:
-    """Test `aswe.api.sports.basketball.get_team_id`"""
+    """Test `aswe.api.sport.basketball.get_team_id`"""
     team_id = get_team_id("Denver Nuggets")
 
     assert isinstance(team_id, int)
