+++ tests/api/sports/test_football.py
@@ -1,4 +1,4 @@
-from aswe.api.sports.football import (
+from aswe.api.sport.football import (
     convert_league_name,
     get_current_team_match,
     get_league_standings,
@@ -10,7 +10,7 @@ from aswe.api.sports.football import (
 
 
 def test_convert_league_name() -> None:
-    """Test `aswe.api.sports.football.convert_league_name`"""
+    """Test `aswe.api.sport.football.convert_league_name`"""
     name = convert_league_name("Testing")
 
     assert name is None
@@ -53,7 +53,7 @@ def test_convert_league_name() -> None:
 
 
 def test_get_league_standings() -> None:
-    """Test `aswe.api.sports.football.get_league_standings`"""
+    """Test `aswe.api.sport.football.get_league_standings`"""
     standings = get_league_standings("PL")
 
     assert len(standings) == 20
@@ -64,7 +64,7 @@ def test_get_league_standings() -> None:
 
 
 def test_get_matchday_matches() -> None:
-    """Test `aswe.api.sports.football.get_matchday_matches`"""
+    """Test `aswe.api.sport.football.get_matchday_matches`"""
     matches = get_matchday_matches("FL1", 7)
 
     assert isinstance(matches, list)
@@ -79,7 +79,7 @@ def test_get_matchday_matches() -> None:
 
 
 def test_get_ongoing_matches() -> None:
-    """Test `aswe.api.sports.football.get_matchday_matches`"""
+    """Test `aswe.api.sport.football.get_matchday_matches`"""
     matches = get_ongoing_matches("BL1")
 
     assert isinstance(matches, list)
@@ -95,7 +95,7 @@ def test_get_ongoing_matches() -> None:
 
 
 def test_get_matches_today() -> None:
-    """Test `aswe.api.sports.football.get_matches_today`"""
+    """Test `aswe.api.sport.football.get_matches_today`"""
     matches = get_matches_today("DED")
 
     assert isinstance(matches, list)
@@ -106,7 +106,7 @@ def test_get_matches_today() -> None:
 
 
 def test_get_current_team_match() -> None:
-    """Test `aswe.api.sports.football.get_current_team_match`"""
+    """Test `aswe.api.sport.football.get_current_team_match`"""
     match = get_current_team_match("CL", "Frankfurt")
 
     assert isinstance(match, list)
@@ -121,7 +121,7 @@ def test_get_current_team_match() -> None:
 
 
 def test_get_upcoming_team_matches() -> None:
-    """Test `aswe.api.sports.football.get_upcoming_team_matches`"""
+    """Test `aswe.api.sport.football.get_upcoming_team_matches`"""
     matches = get_upcoming_team_matches("FL1", "PSG", 6)
 
     assert len(matches) == 6
