+++ tests/api/sports/test_handball.py
@@ -1,6 +1,6 @@
 import pytest
 
-from aswe.api.sports.handball import (
+from aswe.api.sport.handball import (
     get_league_id,
     get_league_table,
     get_team_game_today,
@@ -11,7 +11,7 @@ from aswe.utils.error import ApiLimitReached
 
 @pytest.mark.xfail(raises=ApiLimitReached)
 def test_get_league_id() -> None:
-    """Test `aswe.api.sports.handball.get_league_id`"""
+    """Test `aswe.api.sport.handball.get_league_id`"""
     league_id = get_league_id("Bundesliga")
     assert league_id == 39
 
@@ -21,7 +21,7 @@ def test_get_league_id() -> None:
 
 @pytest.mark.xfail(raises=ApiLimitReached)
 def test_get_league_table() -> None:
-    """Test `aswe.api.sports.handball.get_league_table`"""
+    """Test `aswe.api.sport.handball.get_league_table`"""
     standings = get_league_table("Bundesliga")
     assert len(standings) == 18
 
@@ -31,7 +31,7 @@ def test_get_league_table() -> None:
 
 @pytest.mark.xfail(raises=ApiLimitReached)
 def test_get_team_id() -> None:
-    """Test `aswe.api.sports.handball.get_team_id`"""
+    """Test `aswe.api.sport.handball.get_team_id`"""
     team_id = get_team_id("Fuchse Berlin")
     assert team_id == 315
 
@@ -41,7 +41,7 @@ def test_get_team_id() -> None:
 
 @pytest.mark.xfail(raises=ApiLimitReached)
 def test_get_team_game_today() -> None:
-    """Test `aswe.api.sports.handball.get_team_game_today`"""
+    """Test `aswe.api.sport.handball.get_team_game_today`"""
     game = get_team_game_today("Fuchse Berlin")
     assert isinstance(game, list)
